FIX: SOCK_DGRAM implementation issues
Currently there are a few implementation issues preventing use of pingtop
without root privs (or access to raw sockets one way or another).

- receive_one_ping function assumes that the received packet will have
  the ip header. This will not be the case when using SOCK_DGRAM.
- receive_one_ping and ping_once incorrectly assume that packet_id will
  be treated the same by SOCK_RAW and SOCK_DGRAM. When using SOCK_DGRAM,
  the ICMP header will be validated by the networking stack and the set
  packet_id will be overwritten with the local source port instead.
- ping_once checks for SOCK_RAW privs using the uid. This approach will
  fail for when using CAP_NET_RAW and possibly SUID.
