+++ src/pingtop/cli.py
@@ -7,12 +7,11 @@ from pathlib import Path
 import click
 
 from pingtop.app import PingTopApp
-from pingtop.engine.raw_icmp import RawIcmpEngine
+from pingtop.engine.icmp import IcmpEngine
 from pingtop.exporters import export_snapshot
 from pingtop.models import ExportFormat, SessionConfig
 from pingtop.session import PingSession, infer_export_format
 from pingtop.summary import render_summary
-
 
 def _configure_logging(log_level: str, log_file: str | None) -> None:
     level = getattr(logging, log_level.upper(), logging.INFO)
@@ -23,7 +22,6 @@ def _configure_logging(log_level: str, log_file: str | None) -> None:
         format="%(asctime)s %(levelname)s %(name)s %(message)s",
     )
 
-
 def _read_hosts_file(path: str | None) -> list[str]:
     if not path:
         return []
@@ -33,7 +31,6 @@ def _read_hosts_file(path: str | None) -> list[str]:
         for line in file_path.read_text(encoding="utf-8").splitlines()
         if line.strip() and not line.strip().startswith("#")
     ]
-
 
 def _merge_hosts(hosts: tuple[str, ...], hosts_file: str | None) -> list[str]:
     merged: list[str] = []
@@ -47,7 +44,6 @@ def _merge_hosts(hosts: tuple[str, ...], hosts_file: str | None) -> list[str]:
             merged.append(expanded_host)
     return merged
 
-
 def _expand_host(host: str) -> list[str]:
     clean_host = host.strip()
     if not clean_host:
@@ -59,7 +55,6 @@ def _expand_host(host: str) -> list[str]:
     except ValueError as exc:
         raise click.ClickException(f"Invalid network or host: {clean_host}") from exc
     return [str(address) for address in network.hosts()]
-
 
 @click.command(context_settings={"help_option_names": ["-h", "--help"]})
 @click.argument("hosts", nargs=-1)
@@ -121,7 +116,7 @@ def main(
         log_level=log_level,
     )
     session = PingSession(config=config, targets=merged_hosts)
-    app = PingTopApp(session=session, engine=RawIcmpEngine())
+    app = PingTopApp(session=session, engine=IcmpEngine())
     app.run()
 
     snapshot = session.snapshot()
